requirements.txt
Django==1.11
djangorestframework==3.6.3
django-colorfield==0.1.14